% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Costa}
\alias{Costa}
\title{Costa}
\usage{
Costa(size_p, gr_p, Kp, inter_pP, size_P, gr_P, KP, inter_Pp, n)
}
\arguments{
\item{size_p}{number of prey at t0}

\item{gr_p}{prey growth rate}

\item{Kp}{capacity for prey}

\item{inter_pP}{interaction of pred on prey}

\item{size_P}{number of predator at t0}

\item{gr_P}{pred growth rate}

\item{KP}{capacity for pred}

\item{inter_Pp}{interaction of prey on pred}

\item{n}{number of times}
}
\description{
Using the Volterra class with Rcpp
}
