% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IslaNublar.R
\name{lotka}
\alias{lotka}
\title{Lotka-Voltera}
\usage{
lotka(
  prey_t,
  pred_t,
  prey_gr,
  pred_gr,
  prey_K,
  pred_K,
  int_predprey,
  int_preypred,
  n,
  verbose = FALSE
)
}
\arguments{
\item{prey_t}{number of prey at t0}

\item{pred_t}{number of predator at t0}

\item{prey_gr}{prey growth rate}

\item{pred_gr}{pred growth rate}

\item{prey_K}{capacity for prey}

\item{pred_K}{capacity for pred}

\item{int_predprey}{interaction of pred on prey}

\item{int_preypred}{interaction of prey on pred}

\item{n}{number of times}
}
\description{
Lotka-Voltera
}
