% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sorna}
\alias{Sorna}
\title{Costa}
\usage{
Sorna(prey, pred, n, verbose)
}
\arguments{
\item{prey}{list of initial size, gr, K and interactivity}

\item{pred}{list of initial size, gr, K and interactivity}

\item{n}{number of times}
}
\description{
Using the Volterra class with Rcpp
}
